% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-id-functions.R
\name{asset_rank_pct}
\alias{asset_rank_pct}
\title{Asset Rank (Pct)}
\usage{
asset_rank_pct(metric_df, subject, metric_name = NULL)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
value (dbl), time (date or dttm), asset_class (chr), and name (chr).}

\item{subject}{The specific asset ID whose rank will be calculated,
and for whom the sentence will be about.}
}
\value{
A sentence denoting what rank in the asset class the specified asset ID
(aka subject) has in percent form; in other words, what percent of the asset class
had a lower value.
}
\description{
In a table with multiple asset IDs,
returns the percent rank of the specified ID (subject) for the metric in the table.
}
\examples{
\dontrun{
asset_rank_raw(metric_df)
}

}
