% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{change_over_time_random}
\alias{change_over_time_random}
\title{Change Over Time Random}
\usage{
change_over_time_random(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns:
asset_id, name, time, and value.}
}
\value{
A sentence reporting the difference between the value of the
most recent record and that of the randomly selected earlier record.
}
\description{
Reports the change in metric from the most recent record
relative to a randomly selected earlier record in the table submitted.
}
\examples{
\dontrun{
change_over_time_random(metric_df)
}

}
