% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{compare_two_metrics_latest}
\alias{compare_two_metrics_latest}
\title{Compare Two Metrics From Two Assets}
\usage{
compare_two_metrics_latest(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
asset_id_compare (chr), value (dbl), time (date or dttm),
value_compare (dbl), and name (chr).}
}
\value{
A sentence denoting which metric in the submitted dataframe
the given asset id had the lowest rank for.
}
\description{
How do two metricsm, one from one asset and one from another, compare?
}
\examples{
\dontrun{
compare_two_metrics_latest(metric_df)
}

}
