% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{compare_two_metrics_trend}
\alias{compare_two_metrics_trend}
\title{Compare Two Trends}
\usage{
compare_two_metrics_trend(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
asset_id2 (chr),time (date or dttm),value (dbl),value2 (dbl)}
}
\value{
A sentence denoting the difference or similarity between the
trend of two metrics.
}
\description{
How do the timeseries trends amongst two metrics compare?
}
\examples{
\dontrun{
compare_two_metrics_trend(metric_df)
}

}
