% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{describe_trend}
\alias{describe_trend}
\title{Describe Trend}
\usage{
describe_trend(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns:
asset_id, name, time, and value.}
}
\value{
A sentence describing the trend over time of the metric
in the table.
}
\description{
Describes the trend in the data set over time.
if not, returns an empty character string.
}
\examples{
\dontrun{
two_sigma_sentence(metric_df)
}

}
