% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-functions.R
\name{get_price_chart}
\alias{get_price_chart}
\title{Get Price Chart}
\usage{
get_price_chart(symbol, theme)
}
\arguments{
\item{symbol}{(chr) A ticker for a financial asset.}

\item{theme}{(chr) The theme that should be applied to the candlestick chart.}
}
\value{
A link to an image showing a visualization of the metric included.
}
\description{
Return a candlestick chart for a given asset, with the
specified theme.
}
\examples{
\dontrun{
get_price_chart('amzn','stocknews')
}


}
