% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{latest_best_rank}
\alias{latest_best_rank}
\title{Latest Best Rank}
\usage{
latest_best_rank(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
value (dbl), time (date or dttm), and name (chr). NOTE: value should be a
percentile rank between 0 and 100.}
}
\value{
A sentence denoting which metric the given asset_id has the highest
rank for within its asset class.
}
\description{
Out of a series of metrics, which one does the given asset_id
rank best in?
}
\examples{
\dontrun{
latest_best_rank(metric_df)
}

}
