% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{latest_lowest_rank}
\alias{latest_lowest_rank}
\title{Latest Lowest Rank}
\usage{
latest_lowest_rank(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
asset_class (chr), value (dbl), time (date or dttm), and name (chr). NOTE: the value column
should reflect a percentage rank whose value is between 0 (worst rank)
and 100 (best rank).}
}
\value{
A sentence denoting which metric in the submitted dataframe
the given asset id had the lowest rank for.
}
\description{
Out of a series of ranked metrics,
which one does the given asset_id
report the lowest rank in?
}
\examples{
\dontrun{
latest_lowest_rank(metric_df)
}

}
