% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-table-functions.R
\name{make_html_table}
\alias{make_html_table}
\title{Transform a dataframe into an HTML table}
\usage{
make_html_table(metric_df)
}
\arguments{
\item{metric_df}{A dataframe}
}
\value{
An HTML formatted character string ready for insertion into
an HTML document.
}
\description{
Turn a dataframe into an HTML formatted table.
}
\examples{
\dontrun{
make_html_table(metric_df)
}

}
