% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-id-functions.R
\name{most_correlated}
\alias{most_correlated}
\title{Most Correlated}
\usage{
most_correlated(metric_df, subject, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
value (dbl), time (date or dttm)}

\item{subject}{The asset ID to measure correlations for.}

\item{template_id}{Optional. The template to be used for the article.}
}
\value{
A sentence denoting the asset ID most correlated to the
declared subject.
}
\description{
In a table with multiple asset IDs,
reports the asset ID most correlated to the subject asset_Id ID.
}
\examples{
\dontrun{
most_correlated(metric_df)
}

}
