% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-functions.R
\name{one_metric_over_time_story}
\alias{one_metric_over_time_story}
\title{Visualize a single metric (story format)}
\usage{
one_metric_over_time_story(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
time (date or dttm), value (dbl), name (chr)}
}
\value{
A link to an image showing a visualization of the metric included.
}
\description{
Show a metric plotted over time as a line chart, in a format
friendly for Google Web Stories.
}
\examples{
\dontrun{
return_3news_snippets(metric_df)
}

}
