% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{random_fact_3bullets}
\alias{random_fact_3bullets}
\title{Return Three Random Facts}
\usage{
random_fact_3bullets(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
time (date or dttm)}
}
\value{
A sentence denoting which metric in the submitted dataframe
the given asset id had the lowest rank for.
}
\description{
What are three random facts about the data presented?
}
\examples{
\dontrun{
random_fact_3bullets(metric_df)
}

}
