% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{report_3metrics_random}
\alias{report_3metrics_random}
\title{Report Three Random Metrics}
\usage{
report_3metrics_random(metric_df)
}
\arguments{
\item{metric_df}{A dataframe an asset_id column (chr), and an unspecified
number of numeric columns. The column names of the numeric columns will be
used in the generated text.}
}
\value{
Three sentences, each wrapped in HTML bullet tags (<li>), reporting
the value of three randomly selected numeric columns.
}
\description{
What are three random metrics from the data presented?
}
\examples{
\dontrun{
report_3metrics_random(metric_df)
}

}
