% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-functions.R
\name{report_avg_by_rare_factors}
\alias{report_avg_by_rare_factors}
\title{Report Average by Rare Factors}
\usage{
report_avg_by_rare_factors(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe}

\item{template_id}{Optional. A character string.}
}
\value{
A sentence ready for insertion into
an HTML document.
}
\description{
Report average value of factors
 (categorical variables) that are rare.
}
\examples{
\dontrun{
report_avg_by_rare_factors(metric_df)
}

}
