% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-class-functions.R
\name{report_big_moves_in_class}
\alias{report_big_moves_in_class}
\title{Report Big Moves in Class}
\usage{
report_big_moves_in_class(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns: asset_id,
asset_class, name, time, value.}
}
\value{
A sentence denoting which, if any, asset IDs had large moves.
}
\description{
Which asset IDs in a given asset class had a big move
since their prior recorded value?
}
\examples{
\dontrun{
report_big_moves_in_class(metric_df)
}

}
