% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-functions.R
\name{report_factor_and_metric}
\alias{report_factor_and_metric}
\title{Factor and Metric}
\usage{
report_factor_and_metric(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe}

\item{template_id}{Optional. A character string.}
}
\value{
A sentence ready for insertion into
an HTML document.
}
\description{
Report a factor and a metric.
}
\examples{
\dontrun{
report_factor_and_metric(metric_df)
}

}
