% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-id-functions.R
\name{report_highest_performer}
\alias{report_highest_performer}
\title{Report Highest Performer}
\usage{
report_highest_performer(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
value (dbl), time (date or dttm), asset_class (chr), and name (chr).}

\item{template_id}{Optional. The template to be used for the article.}
}
\value{
A sentence denoting the highest performer out of a set of
given asset IDs.
}
\description{
In a table with multiple asset IDs,
reports the asset ID (subject) with the highest value.
}
\examples{
\dontrun{
report_highest_performer(metric_df)
}

}
