% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-functions.R
\name{report_least_common_factor}
\alias{report_least_common_factor}
\title{Report Least Common Factor}
\usage{
report_least_common_factor(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe}

\item{template_id}{Optional. A character string.}
}
\value{
A sentence ready for insertion into
an HTML document.
}
\description{
Report which factors (categorical variables) are the
least common.
}
\examples{
\dontrun{
report_least_common_factor(metric_df)
}

}
