% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{report_metric}
\alias{report_metric}
\title{Report Metric}
\usage{
report_metric(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns:
asset_id, name, time, and value.}

\item{template_id}{The template to use. optional.}
}
\value{
A sentence reporting the latest value of a given metric.
}
\description{
Reports the latest value of a given metric.
}
\examples{
\dontrun{
report_metric(metric_df)
}

}
