% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{report_metric_with_percent_change}
\alias{report_metric_with_percent_change}
\title{Report Metric With Percent Change}
\usage{
report_metric_with_percent_change(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns:
asset_id, name, time, and value.}
}
\value{
A sentence reporting the latest value of a given metric along with its
percent change since the prior value.
}
\description{
Reports the latest value of a given metric, and its percent change
since the previous record.
}
\examples{
\dontrun{
report_metric_with_percent_change(metric_df)
}

}
