% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{report_range}
\alias{report_range}
\title{Report Range}
\usage{
report_range(metric_df, template_id = NULL)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns:
asset_id, name, and value.}

\item{template_id}{Optional. The template to be used for the article.}
}
\value{
A sentence reporting the range of a given metric
in the submitted dataframe.
}
\description{
Reports the range of a given metric.
}
\examples{
\dontrun{
report_range(metric_df)
}

}
