% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-class-functions.R
\name{report_rank}
\alias{report_rank}
\title{Report Rank}
\usage{
report_rank(metric_df, template_id = NULL, subject)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns: asset_id,
asset_class, name, value.
#' @param template_id The template to use. Optional.}
}
\value{
A sentence denoting the rank of a given asset for the metric in question.
}
\description{
What is the rank of a given asset for a given metric for its
stated asset class?
}
\examples{
\dontrun{
report_rank(metric_df)
}

}
