% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-class-functions.R
\name{report_top3_movers_in_class}
\alias{report_top3_movers_in_class}
\title{Report Top 3 Movers in Class}
\usage{
report_top3_movers_in_class(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns: asset_id,
asset_class, name, time, value.}
}
\value{
A sentence denoting up to 3 largest movers on a percentage basis
in a given asset class.
}
\description{
Which 3 asset IDs in a given asset class had the largest
change, on a percentage basis, since their prior recorded value?
}
\examples{
\dontrun{
report_top3_movers_in_class(metric_df)
}

}
