% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news-functions.R
\name{return_3news_snippets}
\alias{return_3news_snippets}
\title{Return 3 News Snippets}
\usage{
return_3news_snippets(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: link (chr),
title (chr), time (date or dttm), author (chr), and description (chr).}
}
\value{
HTML formatted text with links and descriptions of articles.
}
\description{
In a table of news stories, return 3 based on recency
and length.
}
\examples{
\dontrun{
return_3news_snippets(metric_df)
}

}
