% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-metric-functions.R
\name{strongest_trend}
\alias{strongest_trend}
\title{Strongest Trend}
\usage{
strongest_trend(metric_df)
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
asset_class(chr), value (dbl), time (date or dttm), and name (chr).}
}
\value{
A sentence denoting which metric in the table the asset id
in question exhibited the strongest trend in over time.
}
\description{
Out of a series of metrics,
which one does the given asset_id
show the clearest trend in over time?
}
\examples{
\dontrun{
strongest_trend(metric_df)
}

}
