% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-functions.R
\name{timeseries_plot}
\alias{timeseries_plot}
\title{Visualize a single metric}
\usage{
timeseries_plot(metric_df, theme = "default")
}
\arguments{
\item{metric_df}{A dataframe with the following columns: asset_id (chr),
time (date or dttm), value (dbl), name (chr)}
}
\value{
A link to an image showing a visualization of the metric included.

A link a line plot, themed.
}
\description{
Show a metric plotted over time as a line chart.
}
\examples{
\dontrun{
return_3news_snippets(metric_df)
}

}
