% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{two_sigma_sentence}
\alias{two_sigma_sentence}
\title{Two Sigma Event}
\usage{
two_sigma_sentence(metric_df)
}
\arguments{
\item{metric_df}{}
}
\value{
A sentence if a two sigma event has occurred; an empty
character string if not.
}
\description{
Checks if a two sigma change has occurred. If so, returns a sentence,
if not, returns an empty character string.
}
\examples{
\dontrun{
two_sigma_sentence(metric_df)
}

}
