% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-id-functions.R
\name{two_sigma_sentence}
\alias{two_sigma_sentence}
\title{Two Sigma Change}
\usage{
two_sigma_sentence(metric_df)
}
\arguments{
\item{metric_df}{A dataframe requiring the following columns: asset_id,
name, time, and value.}
}
\value{
A sentence if a two sigma event has occurred; an empty
character string if not.
}
\description{
Was there a two sigma change in the most recent event?
}
\examples{
\dontrun{
two_sigma_sentence(metric_df)
}

}
